# midterm/openapi.yaml
openapi: 3.0.3
info:
  title: Course Task Tracker API
  version: 1.0.0
  description: API for managing course tasks stored in memory.

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/tasks:
    get:
      summary: Return all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
            example:
              title: Finish lab5
              course: CS553
              completed: false
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"

  /api/tasks/{id}:
    get:
      summary: Return one task by id
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Replace an existing task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
            example:
              title: Complete checkin two
              course: CS553
              completed: false
      responses:
        "200":
          description: Task replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update part of an existing task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
            example:
              completed: true
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete an existing task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Task deleted
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      description: Task id
      schema:
        type: string
        example: "1"

  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: title cant be an empty string

    NotFound:
      description: Task not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Task not found

  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    Task:
      type: object
      required:
        - id
        - title
        - course
        - completed
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          minLength: 1
          example: Watch Week 5 lecture
        course:
          type: string
          minLength: 1
          example: CS553
        completed:
          type: boolean
          example: false

    TaskInput:
      type: object
      required:
        - title
        - course
        - completed
      properties:
        title:
          type: string
          minLength: 1
          example: Submit midterm
        course:
          type: string
          minLength: 1
          example: CS553
        completed:
          type: boolean
          example: false

    TaskPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: Finish graduate checkin
        course:
          type: string
          minLength: 1
          example: CS553
        completed:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Task not found